lib: increase the revert timeout to 6 seconds
'done' event needs to take a write lock of the module.
If we're in the middle of processing time-consuming operational callback,
which takes a read lock, 'done' event can't be issued after 'change' event
in a timely manner because sysrepo can't take the write lock because of the
read lock.
2 seconds for the revert timeout is too short under this situation.

When sysrepo tries to take the write lock, it calls
sr_shmmod_modinfo_rdlock_upgrade(), whose timeout value is 5 seconds.

This means 6 seconds for the revert timeout should be enough to give up
waiting for the 'done' or 'abort' events.

This commit also makes this timeout configurable.

Signed-off-by: Wataru Ishida <wataru.ishid@gmail.com>